seperate the payload generator out of covid_api_request.js
