+++ Firebase Project/functions/index.js
@@ -1,6 +1,7 @@
 const functions = require("firebase-functions")
 const request = require("request-promise")
 const requester = require("./covid_api_request")
+const payload_generator = require("./payload_script")
 
 exports.webhook = functions.https.onRequest(async(req, res) => {
     const lah = require("lineapihelper")
@@ -13,44 +14,36 @@ exports.webhook = functions.https.onRequest(async(req, res) => {
     if(requester.overview_command(input_text)){
         // out_text = input_text + " cases in Thailand: " + await requester.today_api(input_text) + " Persons.";
         // payload = requester.payload_in_text(out_text)
-        payload = await requester.flex_thailand(input_text)
+        payload = await payload_generator.flex_thailand(input_text)
     }else if(requester.province(input_text)){
         // out_text = input_text + ", Thailand total cases: " + await requester.today_api2(input_text) + " Persons.";
         // payload = requester.payload_in_text(out_text)
-        payload = await requester.flex_province(input_text)
+        payload = await payload_generator.flex_province(input_text)
     }
     else if(requester.world_command(input_text)){
         // out_text = input_text + ": " + await requester.world_api(input_text) + " cases.";
         // payload = requester.payload_in_text(out_text)
-        payload = await requester.flex_world(input_text)
+        payload = await payload_generator.flex_world(input_text)
     }
     else if(input_text === "BOOM"){
         out_text = "God Boom, Welcome Sir!! Have a nice day.";
-        payload = requester.payload_in_text(out_text)
+        payload = payload_generator.payload_in_text(out_text)
     }
     else if(input_text === "Contact"){
         out_text = "contact FB: Puvana Swatvanith";
-        payload = requester.payload_in_text(out_text)
+        payload = payload_generator.payload_in_text(out_text)
     }
     else if(input_text === "Collaborator"){
         out_text = "Collaborator:" + "\n" + "lisbono2001" + "\n" + "Noboomta" + "\n" + "Bhatara007" + "\n" + "toey10112";
-        payload = requester.payload_collaborator(out_text);
+        payload = payload_generator.payload_collaborator(out_text);
     }
     else if(input_text === "help" || input_text === "Help"){
-        payload = requester.payload_help()
+        payload = payload_generator.payload_help()
     }
     else{
         out_text = "No command, try 'help' for more guide." ;
-        payload = requester.payload_in_text(out_text)
+        payload = payload_generator.payload_in_text(out_text)
     }
-
-
-    // var payload = [
-    //     {
-    //       type: "text",
-    //       text: out_text,
-    //     }
-    // ]
 
     return lah.reply(lah.replyToken(), payload)
